+++ src/parse/control_point.rs
@@ -3,8 +3,8 @@ use std::cmp::Ordering;
 /// New rhythm speed change.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct TimingPoint {
-    pub beat_len: f32,
-    pub time: f32,
+    pub beat_len: f64,
+    pub time: f64,
 }
 
 impl PartialOrd for TimingPoint {
@@ -16,8 +16,8 @@ impl PartialOrd for TimingPoint {
 /// [`TimingPoint`](crate::parse::TimingPoint) that depends on a previous one.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct DifficultyPoint {
-    pub time: f32,
-    pub speed_multiplier: f32,
+    pub time: f64,
+    pub speed_multiplier: f64,
 }
 
 impl PartialOrd for DifficultyPoint {
